Refactor: Two equal interfaces becomes one. Fix #1
