+++ src/Two/AbstractProvider.php
@@ -8,9 +8,9 @@ use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use GuzzleHttp\ClientInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
-use Laravel\Socialite\Contracts\Provider as ProviderContract;
+use Laravel\Socialite\Contracts\ProviderInterface;
 
-abstract class AbstractProvider implements ProviderContract
+abstract class AbstractProvider implements ProviderInterface
 {
     /**
      * The HTTP request instance.
