+++ src/Two/GoogleProvider.php
@@ -3,6 +3,7 @@
 namespace Laravel\Socialite\Two;
 
 use Illuminate\Support\Arr;
+use Laravel\Socialite\Contracts\ProviderInterface;
 
 class GoogleProvider extends AbstractProvider implements ProviderInterface
 {
